fix conv_transpose unittest. test=develop
